Added update function and delete verifications 👾

+++ lib/Screens/UserScreen.dart
@@ -11,7 +11,9 @@ class Userscreen extends StatefulWidget {
   @override
   UserState createState() => UserState();
 
-  const Userscreen({super.key});
+  const Userscreen({super.key, required this.username});
+
+  final String username;
 }
 
 class UserState extends State<Userscreen> {
@@ -20,10 +22,12 @@ class UserState extends State<Userscreen> {
   final TextEditingController _searchController = TextEditingController();
   final RetrieveController _retrieveController = RetrieveController();
   bool _isloading = true;
+  late String username;
 
   @override
   void initState() {
     super.initState();
+    username = widget.username;
     _fetchUsers();
     _searchController.addListener(_filterUsers);
   }
@@ -58,9 +62,31 @@ class UserState extends State<Userscreen> {
   // This is for delete the user
   Future<void> _deleteUser(String id) async {
     try {
-      await FirebaseFirestore.instance.collection('Users').doc(id).delete();
-      _showToast('User deleted successfully', Colors.green);
-      _fetchUsers(); // Refresh the list
+      QuerySnapshot querySnapshot = await FirebaseFirestore.instance.collection('Users').where('username', isEqualTo: username).get();
+      if (querySnapshot.docs.isNotEmpty) {
+        var userdoc = querySnapshot.docs.first;
+        var user = userdoc['username'];
+
+        if (user == username) {
+          _showToast('Cannot delete yourself', Colors.red);
+          return;
+        }
+        else{
+          await FirebaseFirestore.instance.collection('Users').doc(id).delete();
+          _showToast('User deleted successfully', Colors.green);
+          _fetchUsers(); // Refresh the list
+        }
+      }else{
+        Fluttertoast.showToast(
+          msg: 'User not found',
+          toastLength: Toast.LENGTH_SHORT,
+          gravity: ToastGravity.BOTTOM,
+          timeInSecForIosWeb: 1,
+          backgroundColor: Colors.red,
+          textColor: Colors.white,
+          fontSize: 16.0
+        );
+      }
     } catch (e) {
       print('Error deleting user: $e');
       _showToast('Error deleting user', Colors.red);
